### 2021-08-06 #3585 @arfon
@andr1976 – your author entry is missing an affiliation, and your affiliation missing an index. Please check out of the example paper to see how this YAML should be structured: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
