+++ joss/paper.bib
@@ -49,14 +49,6 @@ conditions",
       publisher = "Scandpower Risk Management AS",
 }
  
-@book{cengel,
-   author = "Yunus A. \c{C}engel and Afshin J. Ghajar",
-   year = "2011",
-   title = "Heat and mass transfer - Fundementals and Applications",
-   publisher = "McGraw-Hill",
-   edition = "Fourth",
-}
-
 @book{geankoplis,
   title={Transport Processes and Unit Operations},
   author={Geankoplis, C.J.},
@@ -169,17 +161,8 @@ volume = {39},
 number = {16},
 pages = {8495-8509},
 year = {2014},
-doi = {https://doi.org/10.1016/j.ijhydene.2014.03.028},
+doi = {10.1016/j.ijhydene.2014.03.028},
 author = {Michael Striednig and Stefan Brandstätter and Markus Sartory and Manfred Klell},
-}
-
-@inproceedings{Thorncroft2007AC2M,
-  title={AC 2007-2695: MODELING COMPRESSIBLE AIR FLOW IN A CHARGING OR DISCHARGING VESSEL AND ASSESSMENT OF POLYTROPIC EXPONENT},
-  author={G. Thorncroft and J. Patton and Raymond G. Gordon},
-  booktitle = {2007 Annual Conference & Exposition},
-  year={2007},
-  city = {Honolulu, Hawaii},
-  doi = {10.18260/1-2--2911},
 }
 
 @article{doi:10.1021/ie4033999,
@@ -204,13 +187,6 @@ author = {Michael Striednig and Stefan Brandstätter and Markus Sartory and Manf
     url = "https://projekter.aau.dk/projekter/files/213880237/Analysis_of_the_application_and_sizing_of_pressure_safety_valves_for_fire_protection_on_offshore_oil_and_gas_installations._By_PECT10_1_F15.pdf"
   }
 
-  @misc{firstlaw,
-    author = "WikiMedia",
-    url = "https://commons.wikimedia.org/wiki/File:First_law_open_system.svg",
-    note = "Adapted from DOE-HDBK-1012/1-92, Thermodynamics, Heat Transfer, and Fluid Flow, Volume 1 of 3, 2014. U.S. Department of Energy",
-    year = " 2008,",
-  }
-
   @article{Walt:2011:NAS:1957373.1957466,
 	author = {Walt, Stefan van der and Colbert, S. Chris and Varoquaux, Gael},
 	title = {The {NumPy} Array: A Structure for Efficient Numerical Computation},
@@ -223,6 +199,7 @@ author = {Michael Striednig and Stefan Brandstätter and Markus Sartory and Manf
 	issn = {1521-9615},
 	pages = {22--30},
 	numpages = {9},
+	doi = {10.1109/mcse.2011.37},
 } 
 
 @article{harris2020array,
@@ -246,7 +223,7 @@ author = {Michael Striednig and Stefan Brandstätter and Markus Sartory and Manf
   pages={261--272},
   year={2020},
   publisher={Nature Publishing Group},
-  doi={10.1038/s41592-019-0686-2}
+  doi={10.1038/s41592-019-0686-2},
 }
 
 @Article{Hunter:2007,
@@ -258,6 +235,7 @@ author = {Michael Striednig and Stefan Brandstätter and Markus Sartory and Manf
 	Pages     = {90--95},
 	publisher = {IEEE COMPUTER SOC},
 	year      = 2007,
+	doi	  = {10.1109/mcse.2007.55 },
 }
 
 @InProceedings{mckinney-proc-scipy-2010,
[…]
