### 2021-08-07 #3585 @arfon
> @arfon @timtroendle I have just noticed the JOSE sister journal to JOSS. I believe JOSS is a good fit for the submitted paper, however considering the "applied" nature of my code and it's applicability (also for teaching/education), I think it might be justified to question it as "research code" as such and you may argue that JOSE might be an equally good or better fit. Anyway, I still think JOSS is the right choice, but I will leave it for you to decide, and will accept your judment on this. Preferably before the formal review is started.

@andr1976 - thanks for flagging this. I'll query the scope here which will give the rest of the JOSS editorial team a chance to give feedback on this matter. This will take a week or so.
