### 2021-08-20 #3585 @kthyng
Hi @andr1976 — the editorial board is fine with this as a JOSS submission, so I have removed the `query-scope` label.
