### 2021-09-08 6ac5380
Update streamlit_app.py
Remove beta versions
