### 2021-09-08 8cdf438
Update streamlit_simple.py
Remove beta version functions
