### 2021-09-08 08a7839
Update streamlit_sbapp.py
Remove beta functions
