### 2021-09-14 andr1976/HydDown#57 @mefuller
Contributing, reporting, and support
For review openjournals/joss-reviews#3695 :

Other than the default github features (author identification, issue tracker, pull requests), there is no explicit mention in the README or elsewhere as to how one contributes to the software, reports, bugs (and requests enhancements) or asks for support.
Given the nature of the github interface, I do not think this is strictly required, but it would be beneficial to add text on this to the README file
