### 2021-09-14 andr1976/HydDown#56 @andr1976
@mefuller
I suspect the symlink for pip points to a py2.7 installation also present on your system. Maybe try
`
   pip3 install hyddown
   python3 -m pip install hyddown
`
where python3 is the symlink to your python3 executable
