### 2021-09-14 andr1976/HydDown#57 @andr1976
@mefuller, Good point I will create a contributing.md file and populate it appropriately ASAP
