### 2021-09-14 andr1976/HydDown#56 @mefuller
There appears to be a bug in Fedora where `pip`, even when installed from `python3-pip` is defaulting to a python2.7 version,
but I believe the README should still be updated to reflect the need to use python3:

You will notice that while there is a file `pip3`, python2 has been totally deprecated and the `pip` command is exclusively the python3 variant (pip>21.0 should not have python2 support: https://github.com/pypa/pip)

I flied a bug report (https://bugzilla.redhat.com/show_bug.cgi?id=2004147) and have my output below for your review - do with it as you please, but perhaps simply providing `$ python3 -m pip install hyddown` as a workaround in case of a python2-related error would be a good update

```
$ dnf provides pip3

python3-pip-21.0.1-2.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : fedora
Matched from:
Filename    : /usr/bin/pip3

python3-pip-21.0.1-2.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : updates-testing
Matched from:
Filename    : /usr/bin/pip3

python3-pip-21.0.1-3.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : @System
Matched from:
Filename    : /usr/bin/pip3

python3-pip-21.0.1-3.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : updates
Matched from:
Filename    : /usr/bin/pip3

$ dnf provides pip

python3-pip-21.0.1-2.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : fedora
Matched from:
Provide    : pip = 21.0.1-2.fc34

python3-pip-21.0.1-2.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : updates-testing
Matched from:
Provide    : pip = 21.0.1-2.fc34

python3-pip-21.0.1-3.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : @System
Matched from:
Provide    : pip = 21.0.1-3.fc34

python3-pip-21.0.1-3.fc34.noarch : A tool for installing and managing Python3 packages
Repo        : updates
Matched from:
Provide    : pip = 21.0.1-3.fc34

$ python3 -m pip install hyddown
Defaulting to user installation because normal site-packages is not writeable
Collecting hyddown
  Downloading HydDown-0.15.1-py3-none-any.whl (27 kB)
Installing collected packages: hyddown
Successfully installed hyddown-0.15.1
```
