### 2021-09-14 3b1d89c
Update README.md
Revise installation instructions to account for apparent bug in some current pip(3) implementations
