### 2021-09-14 4665472
Update MANUAL.md
Addressing #56
