### 2021-09-14 andr1976/HydDown#59 @mefuller
Usage and examples
While the manual provides additional information about where to find various main programs, some clarification in the README and comments in the code are strongly needed to make the software more intuitive
(Don't do anything yet, I am still working on the manual)
