### 2021-09-19 3e5d58d
fix spelling, grammar, formatting
