+++ docs/MANUAL.md
@@ -11,18 +11,24 @@ listings: True
 ---
 
 # Introduction
-HydDown is an open source python3 tool for calculation of Hydrogen (or other pure gas phase species) vessel/container depressurization and filling. The HydDown logo shown in [@Fig:logo] visualizes the key parameters and transport phenomena during gas vessel filling or discharging. The thermodynamic state inside the vessel changes over time as seen from immediately observable variables temperature (T) and pressure (P). This is caused by change in fluid inventory (density) due to flow of gas either in- or out of the vessel. Further, heat is transfered from or to the surroundings via convective heat transfer on the in- and outside of the vessel - with heat being conducted thorugh the vessel wall. 
+HydDown is an open source python3 tool for calculation of hydrogen (or other pure gas phase species) vessel/container depressurization and filling.
+The HydDown logo shown in [@Fig:logo] visualizes the key parameters and transport phenomena during gas vessel filling or discharging.
+The thermodynamic state inside the vessel changes over time as seen from immediately observable variables temperature (T) and pressure (P).
+This is caused by change in fluid inventory (density) due to flow of gas either in or out of the vessel.
+Further, heat is transferred from or to the surroundings via convective heat transfer on the in- and outside of the vessel with heat being conducted through the vessel wall.
 
 ![HydDown logo](img/Sketch.png){#fig:logo}
 
-Run the code as simple as: 
+Running the code is as simple as: 
 
     python main.py input.yml
 
 where `main.py` is the main script and `input.yml` is the input file in Yaml syntax. 
 
 ## Background
-HydDown started as a small spare time project for calculation of vessel filling and depressurisation behaviour. This was mainly to demonstrate, that although perceived as a very tedious/difficult task to write your own code for such an apparent complex problem, actually a fairly limited amount of code is necessary if you have a good thermodynamic backend. The code has evolved to a point where additional features will increase the complexity to the point where it is no longer a simple tool. 
+HydDown started as a small spare-time project for calculation of vessel filling and depressurization behaviour.
+This was mainly to demonstrate that although perceived as a very tedious and difficult task to write your own code for such an apparently complex problem, actually only a fairly limited amount of code is necessary if you have a good thermodynamic backend.
+The code has evolved to a point where additional features will increase the complexity to the point where it is no longer a simple tool. 
 
 A few choices has been made to keep things simple:
 
@@ -31,9 +37,13 @@ A few choices has been made to keep things simple:
 - Gas phase only
 - No temperature stratification in the gas phase
 - No temperature gradient through vessel wall 
-- Heat transfer is modelled as simple as possible
+- Heat transfer is modelled as constant or simplified using empirical
+  correlations
 
-These choices make the problem a lot more simple to solve. First of all, the the pure substance Helmholtz energy based equation of state (HEOS) in coolprop offers a lot of convenience in terms of the property pairs/state variables that can be set independently. Using only a single gas phase species also means that component balances is redundant and 2 or 3-phase flash calculations are not required. That being said, the principle used for a single component is more or less the same, even for multicomponent mixtures with potentially more than one phase.
+These choices make the problem much simpler to solve:
+First of all, the the pure substance Helmholtz energy based equation of state (HEOS) in `coolprop` offers a lot of convenience in terms of the property pairs/state variables that can be set independently.
+Using only a single gas phase species also means that component balances is redundant and 2 or 3-phase flash calculations are not required.
[…]
