### 2021-09-19 andr1976/HydDown#59 @mefuller
The exact meaning/definition of specified h, constant U, and constant Q heat flux is lacking (manual.md).

I am also making a number of edits to spelling and grammar in the manual, which I will submit as a PR.
