### 2021-09-19 b9ac603
Update MANUAL.md
Adding more description to the different heat transfer methods /addressing #59
