### 2021-09-19 andr1976/HydDown#59 @andr1976
I have tried to address the meaning of specified h/ const U, constant Q in the manual
