### 2021-09-19 6caf8ea
Create manual_update.yml
Adding action to compile manual on new commits
