### 2021-09-19 32e7540
Update manual_update.yml
