### 2021-09-19 fb624b7
Update manual_update.yml
