### 2021-09-19 fe210a4
Update manual_update.yml
