+++ .github/workflows/manual_update.yml
@@ -7,7 +7,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
-      - run: cd docs/
+      - run: cd docs/ && mkdir output
       - uses: docker://pandoc/core
         with:
           args: >-  # allows you to break string into multiple lines
