+++ .github/workflows/manual_update.yml
@@ -20,8 +20,7 @@ jobs:
             --output=MANUAL.pdf 
             --listings
             MANUAL.md 
-     - name: Upload
-       uses: actions/upload-artifact@v1
+     - uses: actions/upload-artifact@v1
        with:
          name: manual
          # This is the output path where Pandoc will write the compiled
