+++ .github/workflows/manual_update.yml
@@ -20,11 +20,8 @@ jobs:
             --output=MANUAL.pdf 
             --listings
             MANUAL.md 
-     - uses: actions/upload-artifact@v1
-       with:
-         name: manual
-         # This is the output path where Pandoc will write the compiled
-         # PDF. Note, this should be the same directory as the input
-         # paper.md
-         path: docs/paper.pdf
+      - uses: actions/upload-artifact@master
+        with:
+          name: manual
+          path: docs/paper.pdf
      
