### 2021-09-19 ced0526
Update manual_update.yml
