+++ .github/workflows/manual_update.yml
@@ -23,5 +23,5 @@ jobs:
       - uses: actions/upload-artifact@master
         with:
           name: manual
-          path: docs/paper.pdf
+          path: docs/MANUAL.pdf
      
