+++ .github/workflows/manual_update.yml
@@ -20,7 +20,7 @@ jobs:
             --output=MANUAL.pdf 
             --listings
             MANUAL.md 
-      - uses: actions/upload-artifact@master
+      - uses: actions/upload-artifact@main
         with:
           name: manual
           path: docs/MANUAL.pdf
