### 2021-09-19 72421c2
Update manual_update.yml
