### 2021-09-19 478585c
Update manual_update.yml
