+++ .github/workflows/manual_update.yml
@@ -10,7 +10,7 @@ jobs:
       - run: | 
           cd docs
           mkdir output
-      - uses: docker://pandoc/core
+      - uses: docker://pandoc/latex
         with:
           args: "--toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --number-sections --listings --output=output/MANUAL.pdf MANUAL.md"
       - uses: actions/upload-artifact@master
