+++ .github/workflows/manual_update.yml
@@ -12,7 +12,7 @@ jobs:
           mkdir output
       - uses: docker://pandoc/latex
         with:
-          args: "--toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --number-sections --listings --output output/MANUAL.pdf MANUAL.md"
+          args: "--toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --number-sections --listings -o output/MANUAL.pdf -s MANUAL.md"
       #- uses: actions/upload-artifact@master
       #  with:
       #    name: MANUAL
