### 2021-09-19 cf85e75
Update manual_update.yml
