+++ .github/workflows/manual_update.yml
@@ -10,7 +10,7 @@ jobs:
       - run: | 
           cd /home/runner/work/HydDown/HydDown/docs
           mkdir output
-      - uses: docker://pandoc/latex
+      - uses: docker://rstropek/pandoc-latex
         with:
           args: "--toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --number-sections --listings -o /home/runner/work/HydDown/HydDown/output/MANUAL.pdf -s /home/runner/work/HydDown/HydDown/docs/MANUAL.md"
       #- uses: actions/upload-artifact@master
