+++ .github/workflows/manual_update.yml
@@ -12,7 +12,7 @@ jobs:
           mkdir output
       - uses: docker://pandoc/latex
         with:
-          args: "--toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --data-dir=/home/runner/work/HydDown/HydDown/docs/templates/ --resource-path=.:docs --template eisvogel.tex --number-sections --listings -o /home/runner/work/HydDown/HydDown/docs/output/MANUAL.pdf  /home/runner/work/HydDown/HydDown/docs/MANUAL.md"
+          args: "--toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --data-dir=/home/runner/work/HydDown/HydDown/docs/templates/ --resource-path=.:docs --template eisvogel.tex --number-sections --listings -o output/MANUAL.pdf  MANUAL.md"
       #- uses: actions/upload-artifact@master
       #  with:
       #    name: MANUAL
