+++ .github/workflows/manual_update.yml
@@ -10,9 +10,10 @@ jobs:
       - run: | 
           cd /home/runner/work/HydDown/HydDown/docs
           mkdir output
+          echo pwd
       - uses: docker://pandoc/latex
         with:
-          args: --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --data-dir=templates --template eisvogel.tex --number-sections --listings -o output/MANUAL.pdf  MANUAL.md
+          args: --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --data-dir=templates --number-sections --listings -o output/MANUAL.pdf  MANUAL.md
       #- uses: actions/upload-artifact@master
       #  with:
       #    name: MANUAL
