### 2021-09-19 e7c4675
Update manual_update.yml
