### 2021-09-19 1157589
Update manual_update.yml
