### 2021-09-19 5804b64
Update manual_update.yml
