+++ .github/workflows/manual_update.yml
@@ -13,9 +13,9 @@ jobs:
           pwd
       - uses: docker://pandoc/latex
         with:
-          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings -o /home/runner/work/HydDown/HydDown/docs/output/MANUAL.pdf  /home/runner/work/HydDown/HydDown/docs/MANUAL.md
+          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings -o output/MANUAL.pdf  /home/runner/work/HydDown/HydDown/docs/MANUAL.md
       - uses: actions/upload-artifact@master
         with:
           name: MANUAL
-          path: /home/runner/work/HydDown/HydDown/docs/output/MANUAL.pdf
+          path: output/MANUAL.pdf
      
