### 2021-09-19 b887ac0
Update manual_update.yml
