+++ .github/workflows/manual_update.yml
@@ -9,13 +9,13 @@ jobs:
       - uses: actions/checkout@v2
       - run: | 
           cd /home/runner/work/HydDown/HydDown/docs
-          mkdir output
+          mkdir /home/runner/work/HydDown/HydDown/docs/output
           pwd
       - uses: docker://pandoc/latex
         with:
-          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings -o output/MANUAL.pdf  /home/runner/work/HydDown/HydDown/docs/MANUAL.md
+          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings -o /home/runner/work/HydDown/HydDown/docs/output/MANUAL.pdf  /home/runner/work/HydDown/HydDown/docs/MANUAL.md
       - uses: actions/upload-artifact@v2
         with:
           name: MANUAL
-          path: /github/workspace/docs/output/MANUAL.pdf
+          path: /home/runner/work/HydDown/HydDown/docs/output/MANUAL.pdf
      
