### 2021-09-19 e25d477
Update manual_update.yml
