### 2021-09-19 3e14fac
Update manual_update.yml
