### 2021-09-19 c52aa7f
Update manual_update.yml
