### 2021-09-19 1b2770a
Update manual_update.yml
