+++ .github/workflows/manual_update.yml
@@ -8,12 +8,11 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - run: | 
-          cd docs
           mkdir output
           pwd
       - uses: docker://pandoc/latex
         with:
-          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings --output=/home/runner/work/HydDown/HydDown/docs/output/MANUAL.pdf  MANUAL.md
+          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings --output=output/MANUAL.pdf  MANUAL.md
       - run: | 
           ls /home/runner/work/HydDown/HydDown/docs/output/
           echo "docs"
@@ -21,5 +20,5 @@ jobs:
       - uses: actions/upload-artifact@v2
         with:
           name: MANUAL
-          path: /home/runner/work/HydDown/HydDown/docs/output/MANUAL.pdf
+          path: output
      
