+++ .github/workflows/Advanced.yaml
@@ -9,15 +9,10 @@ jobs:
       - uses: actions/checkout@v2
       - name: create file list
         id: files_list
-        run: |
-          echo "Lorem ipsum" > lorem_1.md  # create two example files
-          echo "dolor sit amet" > lorem_2.md
-          mkdir output  # create output dir
-          # this will also include README.md
-          echo "::set-output name=files::$(printf '"%s" ' *.md)"
+        run: mkdir output  # create output dir
       - uses: docker://pandoc/latex:2.9
         with:
-          args: --output=output/result.pdf ${{ steps.files_list.outputs.files }}
+          args: --output=output/result.pdf README.pdf
       - uses: actions/upload-artifact@master
         with:
           name: output
