+++ .github/workflows/Advanced.yaml
@@ -12,9 +12,7 @@ jobs:
         run: mkdir output  # create output dir
       - uses: docker://pandoc/latex:2.9
         with:
-          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --filter pandoc-citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings --output=output/Test.pdf  Test.md
-     
-
+          args: --toc  --top-level-division=chapter  --number-sections --listings --output=output/Test.pdf  Test.md
       - uses: actions/upload-artifact@master
         with:
           name: output
