+++ .github/workflows/Advanced.yaml
@@ -11,8 +11,7 @@ jobs:
         id: files_list
         run: |
           mkdir output  # create output dir
-          sudo apt-get -y install texlive-latex-extra
-      - uses: docker://pandoc/latex
+      - uses: docker://rstropek/pandoc-latex
         with:
           args: --toc --filter pandoc-crossref --citeproc --bibliography=docs/references.bib --top-level-division=chapter  --number-sections --listings  --template docs/eisvogel.tex --output=output/Test.pdf  Test.md 
           #-v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings
