### 2021-09-20 c69c08e
Update and rename Test.md to Manual.md
