+++ .github/workflows/Advanced.yaml
@@ -1,4 +1,4 @@
-name: Advanced Usage
+name: Compile manual
 
 on: push
 
