### 2021-09-20 e44a00a
Delete manual_update.yml
