+++ .github/workflows/manual_update.yml
@@ -1,22 +1,0 @@
-name: Manual
-
-on: push
-
-jobs:
-  convert_via_pandoc:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - run: | 
-          mkdir output
-          pwd
-      - uses: docker://pandoc/latex
-        with:
-          args: -v "/home/runner/work/HydDown/HydDown":"/github/workspace" --toc --filter pandoc-crossref --citeproc --bibliography=references.bib --top-level-division=chapter --resource-path=.:/home/runner/work/HydDown/HydDown/docs --data-dir=templates --number-sections --listings --output=output/MANUAL.pdf  MANUAL.md
-      - run: | 
-          ls output/
-      - uses: actions/upload-artifact@v2
-        with:
-          name: MANUAL
-          path: output
-     
