### 2021-09-20 8c6e8d3
Delete MANUAL.md
