+++ docs/MANUAL.md
@@ -1,1046 +1,0 @@
----
-title: HydDown
-subtitle: User guide and technical reference
-author: Anders Andreasen
-titlepage: true
-toc-own-page: true
-book: true
-reference-section-title: References
-bibliography: references.bib
-listings: True
----
-
-# Introduction
-HydDown is an open source python3 tool for calculation of hydrogen (or other pure gas phase species) vessel/container depressurization and filling.
-The HydDown logo shown in [@Fig:logo] visualizes the key parameters and transport phenomena during gas vessel filling or discharging.
-The thermodynamic state inside the vessel changes over time as seen from immediately observable variables temperature (T) and pressure (P).
-This is caused by change in fluid inventory (density) due to flow of gas either in or out of the vessel.
-Further, heat is transferred from or to the surroundings via convective heat transfer on the in- and outside of the vessel with heat being conducted through the vessel wall.
-
-![HydDown logo](img/Sketch.png){#fig:logo}
-
-Running the code is as simple as: 
-
-    python main.py input.yml
-
-where `main.py` is the main script and `input.yml` is the input file in Yaml syntax. 
-
-## Background
-HydDown started as a small spare-time project for calculation of vessel filling and depressurization behaviour.
-This was mainly to demonstrate that although perceived as a very tedious and difficult task to write your own code for such an apparently complex problem, actually only a fairly limited amount of code is necessary if you have a good thermodynamic backend.
-The code has evolved to a point where additional features will increase the complexity to the point where it is no longer a simple tool. 
-
-A few choices has been made to keep things simple:
-
-- [Coolprop](http://www.coolprop.org/) is used as thermodynamic backend
-- Only pure substances are considered (limited multi-component capabilities are included)
-- Gas phase only
-- No temperature stratification in the gas phase
-- No temperature gradient through vessel wall 
-- Heat transfer is modelled as constant or simplified using empirical
-  correlations
-
-These choices make the problem much simpler to solve:
-First of all, the the pure substance Helmholtz energy based equation of state (HEOS) in `coolprop` offers a lot of convenience in terms of the property pairs/state variables that can be set independently.
-Using only a single gas phase species also means that component balances is redundant and 2 or 3-phase flash calculations are not required.
-That being said, the principle used for a single component is more or less the same, even for multicomponent mixtures with potentially more than one phase.
-
-## Getting the software
-The source code can be obtained either from GitHub (via `git` or via the latest tar-ball release) or via **pip** . No packaged releases have currently been planned for **conda**.  
-
-The main branch is located here:
-
-[`https://github.com/andr1976/HydDown`](https://github.com/andr1976/HydDown)
-
-Installation of latest release via **pip**:
-
-    pip install hyddown
-
-In case `pip` links to a v2.7 of python you will get an error. If so try the following:
-
-    python3 -m pip install hyddown
-
-where python3 is the symlink or full path to the python3 executable installed on your system. 
-
-## Requirements
-
-- [Python](http://www.python.org) (3.8 - at least python3)
-- [Numpy](https://numpy.org/)
-- [matplotlib](https://matplotlib.org/)
-- [Coolprop (6.4.1)](http://www.coolprop.org/)
-- [cerberus](https://docs.python-cerberus.org/en/latest/)
-- [PyYaml](https://pypi.org/project/PyYAML/)
-- [pandas](https://pandas.pydata.org/)
[…]
