### 2021-09-20 fb1b7ae
minor formatting, spelling and clarification to heat transfer
