### 2021-09-20 74c997c
Update Compile_manual.yaml
