### 2021-09-20 7fe7614
Update Compile_manual.yaml
Workflow / gh action for automatic compilation and update of MANUAL hardcopy
