### 2021-09-20 a194294
Update Compile_manual.yaml
