### 2021-09-20 d184f6e
Update Compile_manual.yaml
