### 2021-09-20 bfd59f3
Update Compile_manual.yaml
