+++ .github/workflows/Compile_manual.yaml
@@ -25,8 +25,8 @@ jobs:
       - uses: actions/checkout@v2
       - name: Commit files # commit the output folder
         run: |
-        #  git config --local user.email "test@github.com"
-        #  git config --local user.name "GitHub Action test"
+          git config user.name github-actions
+          git config user.email github-actions@github.com
           git add docs/MANUAL.pdf
           git commit -m "Updated manual"
       - name: Push changes # push the output folder to your repo
