### 2021-09-20 b327c16
Update Compile_manual.yaml
