### 2021-09-20 7620150
Heat transfer revisions on relocated Manual.md
