### 2021-09-21 19c259d
Only trigger manual compilation on changes to file
