### 2021-09-21 andr1976/HydDown#59 @andr1976
@mefuller, noted I have created a separate branch https://github.com/andr1976/HydDown/tree/revise-mefuller to address e.g. comments in the code / doc strings etc. I will let you know when to review
