### 2021-09-21 c4c8730
Update Compile_manual.yaml
Trying to use Eisvogel.tex template via url
