### 2021-09-21 1489f06
Update Compile_manual.yaml
