### 2021-09-23 48e2a50
Additional comment blocks and code documentation
