### 2021-09-28 andr1976/HydDown#63 @bocklund
Installation instructions do not include instructions to install dependencies
The installation instructions in the README say to `pip install hyddown`, but the runtime dependencies are not included.

I was able to find the `requirements.txt` and install them manually, but I would suggest either setting up the distribution so pip can install the dependencies automatically, or include installing the dependencies in the installation instructions.
