### 2021-09-28 andr1976/HydDown#59 @bocklund
I also think it would be nice to have something like a quickstart. The instructions in the README and manual to

> Run the code as simple as:
>
> ```
> python main.py input.yml
> ```

does not work if users install from pip instead of from the source tree, since they don't have access to those files. The instructions could be more clear for users who may not be very experienced and don't know where to find the `input.yml` and `main.py` files.
