### 2021-09-28 andr1976/HydDown#64 @bocklund
Error when running example `input.yml`
Running the `input.yml` at the root directory with the `main.py` script fails with the traceback:

```python
{'calculation': [{'eta': ['unknown field']}]}
Traceback (most recent call last):
  File "main.py", line 21, in <module>
    hdown=HydDown(input)
  File "/Users/brandon/Projects/hyddown-rev/HydDown/src/hyddown/hdclass.py", line 31, in __init__
    self.validate_input()
  File "/Users/brandon/Projects/hyddown-rev/HydDown/src/hyddown/hdclass.py", line 47, in validate_input
    raise ValueError("Error in input file")
ValueError: Error in input file
```

Steps to reproduce:

1. Clone the HydDown repository
2. `cd hyddown`
3. `pip install .`
4. `python src/main.py input.yml`
