### 2021-09-28 andr1976/HydDown#65 @bocklund
Documentation/instructions for running tests
I was able to run `pytest` from the main repo and run 33 tests in `src/hyddown/test_all.py`, however there were no instructions for running these and I'm not sure if I'm testing everything that was intended to be tested. The documentation could use a short statement on the correct procedure for running the automated tests.
