### 2021-09-28 #3695 @bocklund
@andr1976: I finished making my first pass on HyDown. I was able to use the Streamlit app online. It looks great. Your documentation looks very detailed and seems to demonstrate the power of HydDown very well. I have opened a few issues to be fixed. Issues blocking my acceptance in the checklist:

- [ ] **Installation** https://github.com/andr1976/HydDown/issues/63
- [ ] **Functionality** https://github.com/andr1976/HydDown/issues/64
- [ ] **Automated tests** automated tests seem to exist, but instructions for running them do not. Tracked by https://github.com/andr1976/HydDown/issues/65. Validation/manual tests clearly exist, but I was not able to run them (same error as https://github.com/andr1976/HydDown/issues/64)

Non-blocking issues:

- [ ] Improving the usage flow in the documentation for `pip` users https://github.com/andr1976/HydDown/issues/59
