### 2021-09-28 9df184f
Fix for #63, installing dependencies automatically with pip
