### 2021-09-28 704839f
Added guidelines for testing, #65
