### 2021-09-28 33f4823
small addition to running main script
