### 2021-09-28 andr1976/HydDown#63 @andr1976
@bocklund thanks for spotting these inconsistencies. I have not yet uploaded a new package to pypi, but will do that when the review is close to final. However, automatic installation of dependencies can be verified by:

      pip uninstall hyddown cerberus pyyaml

.. just uninstalling a few of the dependencies to have them installed again
and then from the source root folder (using the updated setup cfg)

     pip install .
