### 2021-09-28 #3695 @andr1976
@bocklund  thanks for commenting, really appreciated. I have updated the setup to include dependencies, fixed input.yml, updated manual/documentation, including guidelines for running tests. I wanted to create a separate PR with you as reviewer, however one of my workflows/actions kept pushing to main (by mistake) when pushing to a seprate branch, so it didn't really work as intended. No all the fixes have already been merged to main.
