### 2021-09-30 #3695 @bocklund
@fraukewiese all the acceptance-blocking issues that I identified have been completed and I'm happy to recommend that HydDown be accepted to JOSS.
