### 2021-09-30 andr1976/HydDown#59 @andr1976
@mefuller, let me know if you have more comments I should address or revise the paper/code, also consider closing this issue.
