### 2021-10-02 be221ea
Small fixes with plotting WIP
