### 2021-10-02 d5def2c
Imporve input file validation WIP
