### 2021-10-02 #3695 @andr1976
@fraukewiese , if/when you accept the paper for publication, I would like to make a new release to the python package index, which includes the improvements made by the reviewers, before the paper is published
