### 2021-10-02 2309974
Validation of input file WIP
