+++ src/hyddown/validator.py
@@ -354,33 +354,124 @@ def valve_validation(input):
         : bool 
         True for success, False for failure
     """
+
+    schema_psv = {
+        'initial' : {'required' : True},
+        'calculation' : {'required' : True},
+        'validation' : {'required' : False},
+        'vessel' : {'required' : True},
+        'heat_transfer' : {'required' : False},
+        'valve':{
+            'required' : True,
+            'type': 'dict',
+            'allow_unknown': False,
+            'schema':{
+                'type': {'required' : True,'type': 'string', 'allowed': ['orifice', 'psv', 'controlvalve', 'mdot']},
+                'flow': {'required' : True,'type': 'string', 'allowed': ['discharge', 'filling']},
+                'diameter': {'required' : True,'type': 'number', 'min': 0},
+                'discharge_coef': {'required' : True,'type': 'number', 'min': 0},
+                'set_pressure': {'required' : True,'type': 'number', 'min': 0},
+                'end_pressure': {'type': 'number', 'min': 0},
+                'blowdown': {'required' : True,'type': 'number', 'min': 0, 'max': 1},
+                'back_pressure': {'required' : True,'type': 'number', 'min': 0},
+                'Cv': {'type': 'number', 'min': 0},
+                'mdot': {'type': ['number','list']},
+                'time' : {'type': 'list'},
+            }
+        },
+    }
+
+    schema_orifice = {
+        'initial' : {'required' : True},
+        'calculation' : {'required' : True},
+        'validation' : {'required' : False},
+        'vessel' : {'required' : True},
+        'heat_transfer' : {'required' : False},
+        'valve':{
+            'required' : True,
+            'type': 'dict',
+            'allow_unknown': False,
+            'schema':{
+                'type': {'required' : True,'type': 'string', 'allowed': ['orifice', 'psv', 'controlvalve', 'mdot']},
+                'flow': {'required' : True,'type': 'string', 'allowed': ['discharge', 'filling']},
+                'diameter': {'required' : True,'type': 'number', 'min': 0},
+                'discharge_coef': {'required' : True,'type': 'number', 'min': 0},
+                'set_pressure': {'type': 'number', 'min': 0},
+                'end_pressure': {'type': 'number', 'min': 0},
+                'blowdown': {'type': 'number', 'min': 0, 'max': 1},
+                'back_pressure': {'required' : True,'type': 'number', 'min': 0},
+                'Cv': {'type': 'number', 'min': 0},
+                'mdot': {'type': ['number','list']},
+                'time' : {'type': 'list'},
+            }
+        },
+    }
+
+    schema_control_valve = {
+        'initial' : {'required' : True},
+        'calculation' : {'required' : True},
+        'validation' : {'required' : False},
+        'vessel' : {'required' : True},
+        'heat_transfer' : {'required' : False},
+        'valve':{
+            'required' : True,
+            'type': 'dict',
+            'allow_unknown': False,
+            'schema':{
+                'type': {'required' : True,'type': 'string', 'allowed': ['orifice', 'psv', 'controlvalve', 'mdot']},
+                'flow': {'required' : True,'type': 'string', 'allowed': ['discharge', 'filling']},
+                'back_pressure': {'required' : True,'type': 'number', 'min': 0},
+                'Cv': {'required' : True,'type': 'number', 'min': 0},
+            }
+        },
+    }
+
+    schema_mdot = {
+        'initial' : {'required' : True},
+        'calculation' : {'required' : True},
+        'validation' : {'required' : False},
+        'vessel' : {'required' : True},
+        'heat_transfer' : {'required' : False},
+        'valve':{
+            'required' : True,
+            'type': 'dict',
+            'allow_unknown': False,
[…]
