### 2021-10-03 7c7eb95
Update setup.cfg
