+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = HydDown
-version = 0.15.1
+version = 0.16.1
 author = Anders Andreasen
 author_email = andr1976@gmail.com
 description = Hydrogen (or other pure gas phase species) pressure vessel filling and discharge calculations
