### 2021-10-03 ce69385
Adaptign for tqdm progress bar
