### 2021-10-04 110ff94
Update for tqdm and joss
minor corrections
