+++ scripts/hyddown_main.py
@@ -13,7 +13,7 @@ except:
     import os
     hyddown_path = os.path.join(os.path.abspath(os.path.dirname(__file__)),"..","src")
     sys.path.append(os.path.abspath(hyddown_path))
-    
+
     from hyddown import HydDown
 
 
