+++ docs/references.bib
@@ -200,3 +200,44 @@ author = {Michael Striednig and Stefan Brandstätter and Markus Sartory and Manf
     note = "Adapted from DOE-HDBK-1012/1-92, Thermodynamics, Heat Transfer, and Fluid Flow, Volume 1 of 3, 2014. U.S. Department of Energy",
     year = " 2008,",
   }
+
+  @software{casper_da_costa_luis_2021_5517697,
+  author       = {Casper da Costa-Luis and
+                  Stephen Karl Larroque and
+                  Kyle Altendorf and
+                  Hadrien Mary and
+                  richardsheridan and
+                  Mikhail Korobov and
+                  Noam Yorav-Raphael and
+                  Ivan Ivanov and
+                  Marcel Bargull and
+                  Nishant Rodrigues and
+                  Guangshuo CHEN and
+                  Antony Lee and
+                  Charles Newey and
+                  James and
+                  Joshua Coales and
+                  Martin Zugnoni and
+                  Matthew D. Pagel and
+                  mjstevens777 and
+                  Mikhail Dektyarev and
+                  Alex Rothberg and
+                  Alexander and
+                  Daniel Panteleit and
+                  Fabian Dill and
+                  FichteFoll and
+                  Gregor Sturm and
+                  HeoHeo and
+                  Hugo van Kemenade and
+                  Jack McCracken and
+                  MapleCCC and
+                  Max Nordlund},
+  title        = {{tqdm: A fast, Extensible Progress Bar for Python 
+                   and CLI}},
+  month        = sep,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v4.62.3},
+  doi          = {10.5281/zenodo.5517697},
+  url          = {https://doi.org/10.5281/zenodo.5517697}
+}
