+++ docs/references.bib
@@ -66,7 +66,7 @@ conditions",
 	year = {2017}
 }
 
-@Article{safety4010011,
+@article{safety4010011,
 AUTHOR = {Andreasen, Anders and Borroni, Filippo and Zan Nieto, Marcos and Stegelmann, Carsten and P. Nielsen, Rudi},
 TITLE = {On the Adequacy of API 521 Relief-Valve Sizing Method for Gas-Filled Pressure Vessels Exposed to Fire},
 JOURNAL = {Safety},
@@ -201,7 +201,7 @@ author = {Michael Striednig and Stefan Brandstätter and Markus Sartory and Manf
     year = " 2008,",
   }
 
-  @software{casper_da_costa_luis_2021_5517697,
+  @software{tqdmref,
   author       = {Casper da Costa-Luis and
                   Stephen Karl Larroque and
                   Kyle Altendorf and
