### 2021-10-04 b2ff47c
Update paper.bib
Update ref for joss paper
