### 2021-10-04 7161c6c
tqdm final touch and remove time imports
