### 2021-10-04 #3695 @andr1976
@fraukewiese, I have made a new release (0.16.2 on github) accompanied with an updated PyPi package. The new release includes the improvments suggested by the reviewers (and some fixes here and there)  as well as another proofreding round made by the company department language secretary. So I am good to go from here if you are
