### 2021-10-04 baff536
Added O2 to streamlit app
