### 2021-10-05 #3695 @andr1976
@fraukewiese :

The manuscript has been revised accordingly and the ISBN no has been removed.

The latest tagged release https://github.com/andr1976/HydDown/releases/tag/v0.16.2
Achived here: https://doi.org/10.5281/zenodo.5547917
yesterday, i.e. the small corrections just made to the manuscript are not reflected, but the software is the same. I have checked metadata and removed mefuller as author including an anonynous dummy user that I have used for a workflow/github action.
