### 2021-10-06 #3695 @kyleniemeyer
Hi @andr1976, I am doing some final checks before accepting.

In the third paragraph and in the Acknowledgments section, you have a few instances of consecutive citations. In cases like this, please use the citation format like `[@author1:2001; @author2:2001]` to have them combined in a single citation like "(Author1 et al., 2001; Author2 et al., 2002)".

You also have one or two instances where the `@author:2001` format (i.e., without brackets) would be more appropriate, to put the author name(s) in the text like "Author et al. (2001)". I'm specifically looking at "please refer to (Shafig et al., 2020)."  in the statement of need section and "from (Haque et al., 1992)" in the caption of Figure 2.

Please update these and let me know—thanks!
