### 2021-10-06 #3695 @andr1976
@kyleniemeyer thanks for improving the reference style of the paper, much appreciated. I did a final proof-reading myself, but don't have anything to add/correct.  Also a big thanks to editor @fraukewiese, reviewers @mefuller and @bocklund 👍 you rock!, and another big thanks to @arfon and entire team behind joss - I think this model for reviewing papers could be adapted by other journals (although single/double blind is a challenge though) - it is very constructive and productive.
